requirements.txt
asgiref==3.7.2
Django==5.0.7
gunicorn==22.0.0
packaging==23.2
psycopg2-binary==2.9.10
PyJWT==2.8.0
sqlparse==0.5.0
tzdata==2023.3
